refactor(player-portal): decompose _quarantine/dialog/ (phase 3)
dialog/ pooled a generic confirm modal with two sheet-specific
prompt-handling components. Splitting:

- ConfirmDialog → shared/ui/ (a styled replacement for
  window.confirm; reusable anywhere).
- PromptDialog + PromptQueue → characters/sheet/dialog/ (these
  handle the Foundry bridge prompt-request / dialog-request SSE
  events that fire while the sheet is open — sheet-only by nature).
